Fix price precision by upgrading from float to double
Changes:
- Upgrade all price variables from float (32-bit) to double (64-bit)
- Fixes .0000 and .5000 decimal artifacts in BTC prices
- double provides 15-16 significant digits vs float's 6-7
- No performance impact on ESP32-S3 (has hardware double FPU)
- Update version to V0.99j

This resolves the issue where BTC prices displayed with .0000 or .5000
endings about 50% of the time due to float precision limitations.

+++ include/day_avg.h
@@ -16,9 +16,9 @@ enum DayAvgMode : uint8_t {
 
 // Rolling 24h mean state (5-min buckets)
 void dayAvgRollingReset();
-void dayAvgRollingAdd(time_t sampleUtc, float price);
-bool dayAvgRollingGet(time_t nowUtc, float& outMean);
+void dayAvgRollingAdd(time_t sampleUtc, double price);
+bool dayAvgRollingGet(time_t nowUtc, double& outMean);
 int  dayAvgRollingCount();
 
 // Cycle mean computed from current chart buffer (7pm ET cycle)
-bool dayAvgCycleMean(float& outMean);
+bool dayAvgCycleMean(double& outMean);
